Search foonctionne

+++ app/controllers/bookings_controller.rb
@@ -13,6 +13,7 @@ class BookingsController < ApplicationController
 
   def destroy
     booking = Booking.find(params[:id])
+    authorize booking
     booking.destroy
     redirect_to user_path(current_user)
   end
